create creators class for creator methods
